aider: Add `pure-uuid` to the dependencies in `package.json`.
# Aider chat conversation:

ASSISTANT: The error message indicates that the `pure-uuid` module, which is a dependency of the `svelte-zooming-ui` module, cannot be found when you try to use `svelte-zooming-ui` in another project. This typically happens when a dependency required by a library is not listed in the library's `package.json` under `dependencies`, which means it won't be installed automatically when the library is installed in a project.

To resolve this issue, we need to add `pure-uuid` to the `dependencies` section of the `svelte-zooming-ui` `package.json` file. This will ensure that `pure-uuid` is installed along with `svelte-zooming-ui` when it is added to a project.

Here's the change to `package.json` to add `pure-uuid` to the dependencies:

package.json
```json
<<<<<<< SEARCH
	"dependencies": {
		"just-compare": "^2.3.0"
	}
=======
	"dependencies": {
		"just-compare": "^2.3.0",
		"pure-uuid": "^1.8.1"
	}
>>>>>>> REPLACE
```

After making this change, you should publish a new version of the `svelte-zooming-ui` module so that users can install the updated version with the correct dependencies.