aider: Modify on:click event handler to dispatch the event with an object that has a detail property containing the details.type.
# Aider chat conversation:

USER: handleMenuChoice isnt triggered

ASSISTANT: To ensure that the `handleMenuChoice` function is triggered when a menu item is clicked, we need to dispatch the event correctly. In the current code, the `dispatch` function is being called with the event name `'menu-selection'` and the event detail as `details.type`. However, the `handleMenuChoice` function is expecting an event object with a `detail` property.

We need to modify the `on:click` event handler to dispatch the event with an object that has a `detail` property containing the `details.type`. Here's how to do it:

src/routes/+page.svelte
```svelte
<<<<<<< SEARCH
<div on:click={() => dispatch('menu-selection', details.type)} style="cursor: pointer;">
=======
<div on:click={() => dispatch('menu-selection', { detail: details.type })} style="cursor: pointer;">
>>>>>>> REPLACE
```

This change will ensure that when the `menu-selection` event is dispatched, it includes an object with a `detail` property, which is what the `handleMenuChoice` function expects to receive.